requirements.txt
altgraph==0.17.5
beautifulsoup4==4.14.3
defusedxml==0.7.1
fonttools==4.61.1
fpdf2==2.8.5
greenlet==3.3.1
icalendar==7.0.0
numpy==2.4.2
packaging==26.0
pandas==3.0.0
pillow==12.1.1
playwright==1.58.0
pyee==13.0.0
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
PySide6==6.10.2
PySide6_Addons==6.10.2
PySide6_Essentials==6.10.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
setuptools==82.0.0
shiboken6==6.10.2
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
tzdata==2025.3
